CST/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
all-packages==1.0.1
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.6.2.post1
api==0.0.7
asgiref==3.8.1
attrs==23.2.0
backends==1.6.6
base58==2.1.1
beartype==0.19.0
beautifulsoup4==4.12.2
bip32utils==0.3.post4
bip39==0.0.2
bitarray==2.9.2
bitcoinlib==0.6.15
blessed==1.20.0
blessings==1.7
blinker==1.7.0
blis==1.0.1
blobfile==3.0.0
Brotli==1.1.0
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
construct==2.10.68
construct-typing==0.5.6
contexttimer==0.3.3
contourpy==1.3.1
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.10
cytoolz==0.12.3
defusedxml==0.8.0rc2
Deprecated==1.2.15
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
Django>=4.2.0
django-allauth==65.3.0
django-taggit>=4.0.0
djangorestframework==3.15.1
dnspython==2.4.2
docstring_parser==0.16
docutils==0.20.1
ecdsa==0.19.0
et_xmlfile==2.0.0
fake-useragent==1.5.1
fanstatic==1.4
filelock==3.13.1
fire==0.7.0
Flask==3.0.0
Flask-Cors==5.0.0
fonttools==4.55.2
fpdf2==2.8.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.41
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.32.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.61.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.2
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.10
img2pdf==0.5.1
importlib-metadata==7.0.1
install==1.3.5
instaloader==4.10.3
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jinxed==1.3.0
joblib==1.4.2
js==1.0
jsonalias==0.1.1
kali==0.0.8
keyring==24.3.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
llama_models==0.0.61
llama_stack==0.0.61
llama_stack_client==0.0.61
loguru==0.7.2
lru-dict==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.3
mdurl==0.1.2
mk==2.4.1
mnemonic==0.21
mongo==0.2.0
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mport==0.1.1
multiaddr==0.0.9
multidict==6.0.5
murmurhash==1.0.11
mysql-connector-python==8.4.0
mysqlclient==2.2.4
nano==0.10.0
netaddr==1.3.0
networkx==3.3
nh3==0.2.15
nltk==3.9.1
node==1.2.1
nodejs==0.1.1
nose==1.3.7
npm==0.1.1
Nuitka==2.4.11
numpy==2.1.3
nvm==1.0.21
oauthlib==3.2.2
odict==1.9.0
openai==1.3.6
openpyxl==3.1.5
opt_einsum==3.4.0
optional-django==0.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
pdf2image==1.17.0
pikepdf==9.4.2
pillow==10.3.0
pkg==0.2
pkgconfig==1.5.5
pkginfo==1.9.6
platformdirs==4.0.0
pluggy==1.3.0
plum-dispatch==2.5.4
plumber==1.7
preshed==3.0.9
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.1
psa==0.1.0
pyaml==24.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.21.0
pydantic==2.5.2
pydantic_core==2.14.5
pydyf==0.11.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.1
pymango==0.1.1
pymongo==4.6.1
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.9.0
pyphen==0.15.0
pyproject_hooks==1.0.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
python-socks==2.5.3
python3-openid==3.2.0
pytweening==1.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
readme-renderer==42.0
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scipy==1.14.1
scoop==0.7.2.0
selenium==4.24.0
setuptools==69.0.3
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
snap==0.5
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
solana==0.35.1
solders==0.21.0
sortedcontainers==2.4.0
soupsieve==2.5
source==1.2.0
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
sqlparse==0.5.0
srsly==2.4.8
subprocess-tee==0.4.1
sudo==1.0.0
sympy==1.13.1
termcolor==2.4.0
thinc==8.3.2
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
toolz==0.12.1
tor==1.0.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.1
transformers==4.47.0
trio==0.25.1
trio-websocket==0.11.1
twine==4.0.2
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
varint==1.0.2
vboxapi==1.0
virtualenv==20.25.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==63.0
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.8.0
websockets==12.0
websockets_proxy==0.1.3
Werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.17.0
wsproto==1.2.0
xargs==0.5.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.17.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.1
zope.lifecycleevent==5.0
zope.proxy==5.1
zopfli==0.2.3.post1
zstandard==0.23.0
